Add files via upload
Python file that can make use of "Avian_taxonomy.csv". It allows the user to get a common name form a scientific name vice versa. If the name is not found it returns None. The explanation and the bulk of the program was written by OpenAI's ChatGPT Jan 30 Version. Free Research Preview.